+++ lib/include/pl/core/new_lexer_sm.hpp
@@ -1,0 +1,20 @@
+#pragma once
+
+#include <lexertl/state_machine.hpp>
+
+namespace pl::core {
+
+    namespace {
+
+        enum {
+            eEOF, eNewLine, eKWNamedOpTypeConstIdent,
+            eSingleLineComment, eSingleLineDocComment,
+            eMultiLineCommentOpen, eMultiLineDocCommentOpen, eMultiLineCommentClose,
+            eNumber, eString, eSeparator, eDirective, eDirectiveType, eDirectiveParam,
+            eOperator, eChar
+        };
+    } // anonymous namespace
+
+    void new_lexer_compile(lexertl::state_machine &sm);
+
+} // namespace pl::core
